Fix MSVC warnings in jsscope.cpp

+++ js/src/jsscope.cpp
@@ -973,9 +973,9 @@ JSObject::putProperty(JSContext *cx, jsid id,
 
         shape->rawGetter = getter;
         shape->rawSetter = setter;
-        shape->attrs = attrs;
+        shape->attrs = uint8(attrs);
         shape->flags = flags | Shape::IN_DICTIONARY;
-        shape->shortid = shortid;
+        shape->shortid = int16(shortid);
 
         /*
          * We are done updating shape and lastProp. Now we may need to update
